Initial commit.
This file is intended to be copied as `~/.vim/scripts.vim`, or sourced
by an existing `scripts.vim` file.  The script probes the content of the
file being edited by the external command `file`, and tries to see if
it's a "binary" file.  If it is, it sets the filetype to "binary" and
sets the option `'binary'`.